backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
httpx
supabase
python-dotenv